[fli-iam#2059] Merge VIP API controllers
